+++ frontend-ts/src/components/MovieCard/index.tsx
@@ -9,27 +9,7 @@ import MenuItem from '@mui/material/MenuItem'
 import { styled } from '@mui/material/styles'
 import AddBoxOutlinedIcon from '@mui/icons-material/AddBoxOutlined'
 import CardMenu from '../CardMenu'
-
-type MovieType = {
-  title?: string
-  image?: string
-  releaseDate?: string
-}
-
-export type MovieCardProps = {
-  /** what  the max-width is in number.
-   * @default 250
-   */
-  maxWidth: number
-  /** what is the position in string.
-   * @default relative
-   */
-  position?: 'absolute' | 'relative'
-  movie: MovieType
-  onCardSelect: (movie: MovieType) => void
-  /** possibility to choose a background color */
-  backgroundColor?: string
-}
+import type { MovieCardProps } from '../../types'
 
 const CardInfo = styled(CardContent)(({ theme }: { theme?: Theme }) => ({
   '&:last-child': {
