+++ frontend-ts/src/components/MovieCardSelected/index.tsx
@@ -6,29 +6,7 @@ import CardMedia from '@mui/material/CardMedia'
 import Typography from '@mui/material/Typography'
 import MenuItem from '@mui/material/MenuItem'
 import CardMenu from '../CardMenu'
-
-type Genres = {
-  name: string
-}
-
-type MovieType = {
-  title?: string
-  image?: string
-  releaseDate?: string
-  runtime?: string
-  genres: Genres[]
-  id: string
-}
-
-export type MovieCardSelectedProps = {
-  movie: MovieType
-  display?: 'flex' | 'block'
-  /** what the type of minheight is
-   * @default 164
-   */
-  minHeight?: number
-  onCardDelete: (movie: MovieType) => void
-}
+import type { MovieCardSelectedProps } from '../../types'
 
 export const MovieCardSelected = (props: MovieCardSelectedProps) => {
   const { movie, display = 'flex', minHeight = 164, onCardDelete } = props
